+++ tests/integration/multipass/test_launch.py
@@ -68,6 +68,7 @@ def test_launch(instance_name):
         name=instance_name,
         base_configuration=base_configuration,
         image_name="snapcraft:core22",
+        disk_gb=16,
     )
 
     try:
@@ -89,6 +90,7 @@ def test_launch_existing_instance(core22_instance):
         name=core22_instance.name,
         base_configuration=base_configuration,
         image_name="snapcraft:core22",
+        disk_gb=16,
     )
 
     assert isinstance(instance, multipass.MultipassInstance)
@@ -127,6 +129,7 @@ def test_launch_os_incompatible_instance(core22_instance):
         name=core22_instance.name,
         base_configuration=base_configuration,
         image_name="snapcraft:core22",
+        disk_gb=16,
         auto_clean=True,
     )
 
@@ -153,7 +156,7 @@ def test_launch_instance_config_incompatible_instance(core22_instance):
 
     assert exc_info.value.brief == (
         "Incompatible base detected:"
-        " Expected image compatibility tag 'buildd-base-v3', found 'invalid'."
+        " Expected image compatibility tag 'buildd-base-v5', found 'invalid'."
     )
 
     # Retry with auto_clean=True.
@@ -161,6 +164,7 @@ def test_launch_instance_config_incompatible_instance(core22_instance):
         name=core22_instance.name,
         base_configuration=base_configuration,
         image_name="snapcraft:core22",
+        disk_gb=16,
         auto_clean=True,
     )
 
@@ -174,7 +178,7 @@ def test_launch_instance_not_setup_without_auto_clean(core22_instance):
 
     core22_instance.push_file_io(
         destination=base_configuration._instance_config_path,
-        content=io.BytesIO(b"compatibility_tag: buildd-base-v3\nsetup: false\n"),
+        content=io.BytesIO(b"compatibility_tag: buildd-base-v5\nsetup: false\n"),
         file_mode="0644",
     )
 
@@ -196,7 +200,7 @@ def test_launch_instance_not_setup_with_auto_clean(core22_instance):
 
     core22_instance.push_file_io(
         destination=base_configuration._instance_config_path,
-        content=io.BytesIO(b"compatibility_tag: buildd-base-v3\nsetup: false\n"),
+        content=io.BytesIO(b"compatibility_tag: buildd-base-v5\nsetup: false\n"),
         file_mode="0644",
     )
 
@@ -205,6 +209,7 @@ def test_launch_instance_not_setup_with_auto_clean(core22_instance):
         name=core22_instance.name,
         base_configuration=base_configuration,
         image_name="snapcraft:core22",
+        disk_gb=16,
         auto_clean=True,
     )